chore: move pairing script to pairing mod
